Readded languageconfig.csv, relocated readLanguageConfigFile into another file #32
