+++ scripts/Wikidata crawler/languageconfig.csv
@@ -1,0 +1,6 @@
+langkey;language
+en;English
+de;German
+fr;French
+it;Italian
+es;Spanish
